refactor: even more refactoring of conftest
