+++ tests/data_context/test_data_context_v013.py
@@ -14,6 +14,7 @@ from great_expectations.data_context.types.base import (
     DataContextConfig,
     dataContextConfigSchema,
 )
+from great_expectations.data_context.util import file_relative_path
 from great_expectations.exceptions import ExecutionEngineError
 from great_expectations.execution_engine.pandas_batch_data import PandasBatchData
 from great_expectations.execution_engine.sqlalchemy_batch_data import (
@@ -69,6 +70,42 @@ def basic_data_context_v013_config():
             },
         }
     )
+
+
+@pytest.fixture
+def data_context_with_runtime_sql_datasource_for_testing_get_batch(
+    sa,
+    empty_data_context,
+):
+    context: DataContext = empty_data_context
+    db_file_path: str = file_relative_path(
+        __file__,
+        os.path.join("..", "test_sets", "test_cases_for_sql_data_connector.db"),
+    )
+
+    datasource_config: str = f"""
+        class_name: Datasource
+
+        execution_engine:
+            class_name: SqlAlchemyExecutionEngine
+            connection_string: sqlite:///{db_file_path}
+
+        data_connectors:
+            my_runtime_data_connector:
+                module_name: great_expectations.datasource.data_connector
+                class_name: RuntimeDataConnector
+                batch_identifiers:
+                    - pipeline_stage_name
+                    - airflow_run_id
+    """
+
+    context.test_yaml_config(
+        name="my_runtime_sql_datasource", yaml_config=datasource_config
+    )
+
+    # noinspection PyProtectedMember
+    context._save_project_config()
+    return context
 
 
 def test_ConfigOnlyDataContext_v013__initialization(
