+++ tests/execution_engine/test_sparkdf_execution_engine.py
@@ -133,6 +133,22 @@ def spark_df_from_pandas_df():
     return _construct_spark_df_from_pandas
 
 
+@pytest.fixture
+def test_folder_connection_path_tsv(tmp_path_factory):
+    df1 = pd.DataFrame({"col_1": [1, 2, 3, 4, 5], "col_2": ["a", "b", "c", "d", "e"]})
+    path = str(tmp_path_factory.mktemp("test_folder_connection_path_tsv"))
+    df1.to_csv(path_or_buf=os.path.join(path, "test.tsv"), sep="\t", index=False)
+    return str(path)
+
+
+@pytest.fixture
+def test_folder_connection_path_parquet(tmp_path_factory):
+    df1 = pd.DataFrame({"col_1": [1, 2, 3, 4, 5], "col_2": ["a", "b", "c", "d", "e"]})
+    path = str(tmp_path_factory.mktemp("test_folder_connection_path_parquet"))
+    df1.to_parquet(path=os.path.join(path, "test.parquet"))
+    return str(path)
+
+
 def test_reader_fn(spark_session, basic_spark_df_execution_engine):
     engine = basic_spark_df_execution_engine
     # Testing that can recognize basic csv file
